+++ examples/forest_fire.jl
@@ -43,11 +43,11 @@ function model_initiation(; f = 0.02, d = 0.8, p = 0.01, griddims = (100, 100), 
     properties = Dict(:f => f, :d => d, :p => p)
     forest = AgentBasedModel(Tree, space; properties = properties)
 
-    ## create and add trees to each node with probability d,
+    ## create and add trees to each position with probability d,
     ## which determines the density of the forest
-    for node in nodes(forest)
+    for position in positions(forest)
         if rand() ≤ forest.d
-            add_agent!(node, forest, true)
+            add_agent!(position, forest, true)
         end
     end
     return forest
@@ -60,11 +60,11 @@ forest = model_initiation(f = 0.05, d = 0.8, p = 0.05, griddims = (20, 20), seed
 # stepping function for the model
 
 function forest_step!(forest)
-    for node in nodes(forest, by = :random)
-        np = agents_in_pos(node, forest)
+    for position in positions(forest, by = :random)
+        np = agents_in_pos(position, forest)
         ## the position is empty, maybe a tree grows here
         if length(np) == 0
-            rand() ≤ forest.p && add_agent!(node, forest, true)
+            rand() ≤ forest.p && add_agent!(position, forest, true)
         else
             tree = forest[np[1]] # by definition only 1 agent per position
             if tree.status == false  # if it is has been burning, remove it.
@@ -73,7 +73,7 @@ function forest_step!(forest)
                 if rand() ≤ forest.f  # the tree ignites spontaneously
                     tree.status = false
                 else  # if any neighbor is on fire, set this tree on fire too
-                    for pos in nearby_positions(node, forest)
+                    for pos in nearby_positions(position, forest)
                         neighbors = agents_in_pos(pos, forest)
                         length(neighbors) == 0 && continue
                         if any(n -> !forest.agents[n].status, neighbors)