+++ examples/wealth_distribution.jl
@@ -85,7 +85,7 @@ end
 function wealth_model_2D(; dims = (25, 25), wealth = 1, M = 1000)
     space = GridSpace(dims, periodic = true)
     model = ABM(WealthInSpace, space; scheduler = random_activation)
-    for i in 1:M # add agents in random nodes
+    for i in 1:M # add agents in random positions
         add_agent!(model, wealth)
     end
     return model
@@ -95,13 +95,12 @@ model2D = wealth_model_2D()
 
 # The agent actions are a just a bit more complicated in this example.
 # Now the agents can only give wealth to agents that exist on the same or
-# neighboring nodes (their "neighbors").
+# neighboring positions (their "neighbors").
 
 function agent_step_2d!(agent, model)
     agent.wealth == 0 && return # do nothing
-    agent_node = coord2vertex(agent.pos, model)
-    neighboring_positions = nearby_positions(agent_node, model)
-    push!(neighboring_positions, agent_node) # also consider current position
+    neighboring_positions = nearby_positions(agent.pos, model)
+    push!(neighboring_positions, agent.pos) # also consider current position
     rpos = rand(neighboring_positions) # the position that we will exchange with
     available_ids = agents_in_pos(rpos, model)
     if length(available_ids) > 0