+++ src/spaces/discrete.jl
@@ -3,25 +3,25 @@ This file implements functions shared by all discrete spaces.
 Discrete spaces are by definition spaces with a finite amount of possible positions.
 
 All these functions are granted "for free" to discrete spaces by simply extending:
-- nodes(model)
+- positions(model)
 - agents_in_pos(position, model)
 =#
 
-export nodes, agents_in_pos, empty_positions, random_empty, has_empty_positions
+export positions, agents_in_pos, empty_positions, random_empty, has_empty_positions
 
 """
-    nodes(model::ABM{A, <:DiscreteSpace}) → ns
-Return an iterator over all positions of a model with a discrete space (called nodes).
+    positions(model::ABM{A, <:DiscreteSpace}) → ns
+Return an iterator over all positions of a model with a discrete space.
 
-    nodes(model::ABM{A, <:DiscreteSpace}, by::Symbol) → ns
-Return all positions of a model with a discrete space (called nodes), sorting them
+    positions(model::ABM{A, <:DiscreteSpace}, by::Symbol) → ns
+Return all positions of a model with a discrete space, sorting them
 using the argument `by` which can be:
 * `:random` - randomly sorted
 * `:population` - positions are sorted depending on how many agents they accommodate.
   The more populated positions are first.
 """
-function nodes(model::ABM{<:AbstractAgent,<:DiscreteSpace}, by::Symbol)
-    n = collect(nodes(model))
+function positions(model::ABM{<:AbstractAgent,<:DiscreteSpace}, by::Symbol)
+    n = collect(positions(model))
     itr = reshape(n, length(n))
     if by == :random
         shuffle!(itr)
@@ -35,7 +35,7 @@ end
 
 # TODO: Does this really have to be collecting...?
 function empty_positions(model::ABM{<:AbstractAgent,<:DiscreteSpace})
-    collect(Iterators.filter(i -> length(agents_in_pos(i, model)) == 0, nodes(model)))
+    collect(Iterators.filter(i -> length(agents_in_pos(i, model)) == 0, positions(model)))
 end
 
 """
@@ -72,7 +72,7 @@ export add_agent_single!, fill_space!, move_agent_single!
     add_agent_single!(agent::A, model::ABM{A, <: DiscreteSpace}) → agent
 
 Add the `agent` to a random position in the space while respecting a maximum of one agent
-per node position. This function does nothing if there aren't any empty positions.
+per position. This function does nothing if there aren't any empty positions.
 """
 function add_agent_single!(agent::A, model::ABM{A,<:DiscreteSpace}) where {A<:AbstractAgent}
     position = random_empty(model)
@@ -101,11 +101,11 @@ end
 """
     fill_space!([A ,] model::ABM{A, <:DiscreteSpace}, args...; kwargs...)
     fill_space!([A ,] model::ABM{A, <:DiscreteSpace}, f::Function; kwargs...)
-Add one agent to each node in the model's space. Similarly with [`add_agent!`](@ref),
+Add one agent to each position in the model's space. Similarly with [`add_agent!`](@ref),
 the function creates the necessary agents and
 the `args...; kwargs...` are propagated into agent creation.
 If instead of `args...` a function `f` is provided, then `args = f(pos)` is the result of
-applying `f` where `pos` is each position (tuple for grid, node index for graph).
+applying `f` where `pos` is each position (tuple for grid, index for graph).
 
 An optional first argument is an agent **type** to be created, and targets mixed agent
 models where the agent constructor cannot be deduced (since it is a union).
@@ -138,9 +138,9 @@ function fill_space!(
     args...;
     kwargs...,
 ) where {A<:AbstractAgent,U<:AbstractAgent}
-    for n in nodes(model)
+    for p in positions(model)
         id = nextid(model)
-        add_agent_pos!(A(id, n, args...; kwargs...), model)
+        add_agent_pos!(A(id, p, args...; kwargs...), model)
     end
     return model
 end
@@ -151,10 +151,10 @@ function fill_space!(
     f::Function;
     kwargs...,
 ) where {A<:AbstractAgent,U<:AbstractAgent}
-    for n in nodes(model)
+    for p in positions(model)
         id = nextid(model)
-        args = f(n)
-        add_agent_pos!(A(id, n, args...; kwargs...), model)
+        args = f(p)
+        add_agent_pos!(A(id, p, args...; kwargs...), model)
     end
     return model
 end