+++ test/benchmark/forestFire_defs.jl
@@ -10,32 +10,32 @@ function model_initiation(; f = 0.02, d = 0.8, p = 0.01, griddims=(100,100), see
     properties = Dict(:f => f, :d => d, :p => p)
     forest = AgentBasedModel(Tree, space; properties = properties)
 
-    ## create and add trees to each node with probability d,
+    ## create and add trees to each position with probability d,
     ## which determines the density of the forest
-    for node in nodes(forest)
+    for pos in positions(forest)
         if rand() ≤ forest.d
-            add_agent!(node, forest, true)
+            add_agent!(pos, forest, true)
         end
     end
     return forest
 end
 
 function forest_step!(forest)
-  for node in nodes(forest, by = :random)
-    ap = agents_in_pos(node, forest)
+  for pos in positions(forest, by = :random)
+    ap = agents_in_pos(pos, forest)
     ## the cell is empty, maybe a tree grows here
     if length(ap) == 0
-        rand() ≤ forest.p && add_agent!(node, forest, true)
+        rand() ≤ forest.p && add_agent!(pos, forest, true)
     else
-      tree = forest[ap[1]] # by definition only 1 agent per node
+      tree = forest[ap[1]] # by definition only 1 agent per position
       if tree.status == false  # if it is has been burning, remove it.
         kill_agent!(tree, forest)
       else
         if rand() ≤ forest.f  # the tree ignites spontaneously
           tree.status = false
         else  # if any neighbor is on fire, set this tree on fire too
-          for pos in nearby_positions(node, forest)
-            neighbors = agents_in_pos(pos, forest)
+          for n_pos in nearby_positions(pos, forest)
+            neighbors = agents_in_pos(n_pos, forest)
             length(neighbors) == 0 && continue
             if any(n -> !forest.agents[n].status, neighbors)
               tree.status = false