+++ test/benchmark/schelling_defs.jl
@@ -2,7 +2,7 @@
 mutable struct SchellingAgent <: AbstractAgent
   id::Int # The identifier number of the agent
   pos::Tuple{Int,Int} # The x, y location of the agent
-  mood::Bool # whether the agent is happy in its node. (true = happy)
+  mood::Bool # whether the agent is happy in its position. (true = happy)
   group::Int # The group of the agent,
              # determines mood as it interacts with neighbors
 end
@@ -46,7 +46,7 @@ function agent_step!(agent, model)
 
       # After counting the neighbors, decide whether or not to move the agent.
       # If count_neighbors_same_group is at least the min_to_be_happy, set the
-      # mood to true. Otherwise, move the agent to a random node.
+      # mood to true. Otherwise, move the agent to a random position.
       if count_neighbors_same_group ≥ minhappy
           agent.mood = true
       else